added component and routing
