+++ angular-ui/src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import { UserProfileComponent } from './profile/edit-user-profile/edit-user-prof
 import { BrowseComponent } from './browse/browse.component';
 import { ProfileComponent } from './profile/profile.component';
 import { UserLoginComponent } from './user-login/user-login.component';
+import { MyListComponent } from './my-list/my-list.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'browse', pathMatch: 'full' },
@@ -14,6 +15,7 @@ const routes: Routes = [
   { path: 'password-form', component: PasswordFormComponent },
   { path: 'profile', component: ProfileComponent },
   { path: 'login', component: UserLoginComponent },
+  { path: 'my-list', component: MyListComponent},
 
 ];
