+++ angular-ui/src/app/my-list/my-list.component.html
@@ -1,0 +1,1 @@
+<p>my-list works!</p>
