+++ angular-ui/src/app/my-list/my-list.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-my-list',
+  templateUrl: './my-list.component.html',
+  styleUrls: ['./my-list.component.css']
+})
+export class MyListComponent {
+
+}
