+++ angular-ui/src/app/navbar/navbar.component.html
@@ -22,7 +22,7 @@
           <a class="nav-link" href="#">Pick For Me</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link">My List</a>
+          <a class="nav-link" [routerLink]="'my-list'" routerLinkActive="active">My List</a>
         </li>
       </ul>
       <ul class="nav-item dropdown">
