+++ angular-ui/src/app/profile-view/profile-view.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ReviewsService } from 'src/app/shared/reviews.service';
+import { ReviewsService } from 'src/app/shared/review.service';
 import { UserReview } from 'src/app/shared/user-review.model';
 @Component({
   selector: 'app-profile-view',
