+++ angular-ui/src/app/shared/review.service.spec.ts
@@ -1,21 +1,16 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
-import { UserReview } from './user-review.model';
+import { TestBed } from '@angular/core/testing';
 
-@Injectable({
-  providedIn: 'root',
-})
-export class ReviewsService {
-  private apiUrl = 'http://localhost:8080/api/reviews'; // Assuming your Spring Boot app runs on port 8080
+import { ReviewsService } from './review.service';
 
-  constructor(private http: HttpClient) {}
+describe('ReviewService', () => {
+  let service: ReviewsService;
 
-  getAllReviews(): Observable<UserReview[]> {
-    return this.http.get<UserReview[]>(this.apiUrl);
-  }
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ReviewsService);
+  });
 
-  addReview(review: UserReview): Observable<UserReview> {
-    return this.http.post<UserReview>(this.apiUrl, review);
-  }
-}
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
